requirements.txt
faker
datetime
requests
pandas
fastapi
uvicorn
sqlalchemy 
psycopg2-binary
python-dotenv
tabulate